added fonts

+++ 01-dating-app/src/MyProfile.js
@@ -1,11 +1,12 @@
 import React from 'react'
 import axios from 'axios'
+import CreateProfiles from './CreateProfiles'
 
 const baseURL = 'https://3001-white-impala-sa4c1pjn.ws-us03.gitpod.io'
 
 export default class MyProfile extends React.Component {
     state = {
-        username: '',
+        username: this.props.username,
         user_id: this.props.user_id,
         profile: '',
         editedName: '',
@@ -15,78 +16,91 @@ export default class MyProfile extends React.Component {
         editedDob: '',
         editedInterests: [],
         editedIntroduction: '',
-        editedImage:'',
-        isLoggedIn:this.props.isLoggedIn
+        editedImage: '',
+        isLoggedIn: this.props.isLoggedIn
 
     }
     render() {
-        return (
-            <React.Fragment>
+        if (this.state.isLoggedIn === true) {
+            return (
+                <React.Fragment>
 
-                <div id='signup'>
-                    <div className='m-3 text-left'>
-                        <label className='form-label'>Name</label>
-                        <input className="form-control" type="text" name='editedName' value={this.state.editedName} onChange={this.updateFormFields} />
+                    <div id='signup'>
+                        <div className='m-3 text-left'>
+                            <label className='form-label'>Name</label>
+                            <input className="form-control" type="text" name='editedName' value={this.state.editedName} onChange={this.updateFormFields} />
+                        </div>
+
+                        <div className='m-3 text-left'>
+                            <label className='form-label'>Username</label>
+                            <div className='input-group text-left'>
+                                <span className="input-group-text">@</span>
+                                <input className="form-control" type="text" name='username' value={this.state.username} disabled />
+                            </div>
+                        </div>
+
+                        <div className='m-3 text-left'>
+                            <label className='form-label'>Gender</label>
+                            <div className='form-check'>
+                                <input className='form-check-input' type='radio' value='male' name='editedGender' onChange={this.updateFormFields} checked={this.state.editedGender === 'male'} />
+                                <label className='form-check-label'>Male</label>
+                            </div>
+                            <div className='form-check'>
+                                <input className='form-check-input' type='radio' value='female' name='editedGender' onChange={this.updateFormFields} checked={this.state.editedGender === 'female'} />
+                                <label className='form-check-label'>Female</label>
+                            </div>
+                        </div>
+
+                        <div className='m-3 text-left'>
+                            <label className='form-label'>DOB</label>
+                            <input className="form-control" type="date" name='editedDob' value={this.state.editedDob} onChange={this.updateFormFields} />
+                        </div>
+
+                        <div className='m-3'>
+                            <label className='form-label'>I Enjoy...</label>
+                            <div className='form-check'>
+                                <input className='form-check-input' type='checkbox' name='editedInterests' value='sports' onChange={this.updateInterests} checked={this.state.editedInterests.includes('sports')} />
+                                <label className='form-check-label'>Sports</label>
+                            </div>
+                            <div className='form-check'>
+                                <input className='form-check-input' type='checkbox' name='editedInterests' value='arts' onChange={this.updateInterests} checked={this.state.editedInterests.includes('arts')} />
+                                <label className='form-check-label'>Arts</label>
+                            </div>
+                            <div className='form-check'>
+                                <input className='form-check-input' type='checkbox' name='editedInterests' value='food' onChange={this.updateInterests} checked={this.state.editedInterests.includes('food')} />
+                                <label className='form-check-label'>Food</label>
+                            </div>
+                            <div className='form-check'>
+                                <input className='form-check-input' type='checkbox' name='editedInterests' value='travel' onChange={this.updateInterests} checked={this.state.editedInterests.includes('travel')} />
+                                <label className='form-check-label'>Travel</label>
+                            </div>
+                            <div className='form-check'>
+                                <input className='form-check-input' type='checkbox' name='editedInterests' value='education' onChange={this.updateInterests} checked={this.state.editedInterests.includes('education')} />
+                                <label className='form-check-label'>Education</label>
+                            </div>
+                        </div>
+
+                        <div className='m-3 text-left'>
+                            <label className='form-label'>About Me</label>
+                            <textarea className="form-control" type="text" name='editedIntroduction' value={this.state.editedIntroduction} onChange={this.updateFormFields}></textarea>
+                        </div>
+                        <div className='m-3 text-left'>
+                            <label className='form-label'>Photo</label>
+                            <input className="form-control" type="text" name='editedImage' value={this.state.editedImage} onChange={this.updateFormFields} placeholder='Insert image URL here' ></input>
+                        </div>
+                        <div className='d-flex justify-content-between'>
+                            <button className='btn pinkBtn m-3 px-5' onClick={this.editProfile}>Update</button>
+                            <button className='btn delBtn m-3 px-5' onClick={this.deleteProfile}>Delete</button>
+                        </div>
+
                     </div>
 
-                    <div className='m-3 text-left'>
-                        <label className='form-label'>Gender</label>
-                        <div className='form-check'>
-                            <input className='form-check-input' type='radio' value='male' name='editedGender' onChange={this.updateFormFields} checked={this.state.editedGender === 'male'} />
-                            <label className='form-check-label'>Male</label>
-                        </div>
-                        <div className='form-check'>
-                            <input className='form-check-input' type='radio' value='female' name='editedGender' onChange={this.updateFormFields} checked={this.state.editedGender === 'female'} />
-                            <label className='form-check-label'>Female</label>
-                        </div>
-                    </div>
+                </React.Fragment>
 
-                    <div className='m-3 text-left'>
-                        <label className='form-label'>DOB</label>
-                        <input className="form-control" type="date" name='editedDob' value={this.state.editedDob} onChange={this.updateFormFields} />
-                    </div>
-
-                    <div className='m-3'>
-                        <label className='form-label'>I Enjoy...</label>
-                        <div className='form-check'>
-                            <input className='form-check-input' type='checkbox' name='editedInterests' value='sports' onChange={this.updateInterests} checked={this.state.editedInterests.includes('sports')} />
-                            <label className='form-check-label'>Sports</label>
-                        </div>
-                        <div className='form-check'>
-                            <input className='form-check-input' type='checkbox' name='editedInterests' value='arts' onChange={this.updateInterests} checked={this.state.editedInterests.includes('arts')} />
-                            <label className='form-check-label'>Arts</label>
-                        </div>
-                        <div className='form-check'>
-                            <input className='form-check-input' type='checkbox' name='editedInterests' value='food' onChange={this.updateInterests} checked={this.state.editedInterests.includes('food')} />
-                            <label className='form-check-label'>Food</label>
-                        </div>
-                        <div className='form-check'>
-                            <input className='form-check-input' type='checkbox' name='editedInterests' value='travel' onChange={this.updateInterests} checked={this.state.editedInterests.includes('travel')} />
-                            <label className='form-check-label'>Travel</label>
-                        </div>
-                        <div className='form-check'>
-                            <input className='form-check-input' type='checkbox' name='editedInterests' value='education' onChange={this.updateInterests} checked={this.state.editedInterests.includes('education')} />
-                            <label className='form-check-label'>Education</label>
-                        </div>
-                    </div>
-
-                    <div className='m-3 text-left'>
-                        <label className='form-label'>About Me</label>
-                        <textarea className="form-control" type="text" name='editedIntroduction' value={this.state.editedIntroduction} onChange={this.updateFormFields}></textarea>
-                    </div>
-                    <div className='m-3 text-left'>
-                        <label className='form-label'>Photo</label>
-                        <input className="form-control" type="text" name='editedImage' value={this.state.editedImage} onChange={this.updateFormFields} placeholder='Insert image URL here' ></input>
-                    </div>
-                    <div className='d-flex justify-content-between'>
-                        <button className='btn btn-primary m-3 px-5' onClick={this.editProfile}>Confirm</button>
-                        <button className='btn btn-danger m-3 px-5' onClick={this.deleteProfile}>Delete</button>
-                    </div>
-                    
-                </div>
-
-            </React.Fragment>
-        )
+            )
+        } else {
+            return (<CreateProfiles />)
+        }
     }
 
     updateFormFields = event => {
@@ -107,7 +121,7 @@ export default class MyProfile extends React.Component {
     }
 
     async componentDidMount() {
-        
+        if (this.state.isLoggedIn === true) {
             let foundUser = {
                 user_id: this.state.user_id
 
@@ -122,25 +136,12 @@ export default class MyProfile extends React.Component {
                 editedDob: userProfile.data.dob,
                 editedInterests: userProfile.data.interests,
                 editedIntroduction: userProfile.data.introduction,
-                editedImage:userProfile.data.image,
+                editedImage: userProfile.data.image,
                 validate: false
             })
 
-        
+        }
     }
-
-    // showError = () => {
-    //     let char = ['!', '@', '#', '$', '%', '^', '&', '*']
-    //     for (let x of char) {
-    //         if (this.state.username.includes(x) && this.state.username.length > 4) {
-    //             return true
-    //         } else {
-    //             continue
-    //         }
-
-    //     }
-
-    // }
 
     updateInterests = event => {
         if (this.state.editedInterests.includes(event.target.value) === false) {
@@ -182,12 +183,16 @@ export default class MyProfile extends React.Component {
 
     }
 
-    deleteProfile= event=>{
-        
-        let response=axios.delete(baseURL+'/deleteProfile/' + this.state.user_id)
-        console.log(response)
-        axios.delete(baseURL+'/deleteUsername/'+ this.state.user_id)
-        window.location = '/'
+    deleteProfile = event => {
+        if (window.confirm('Are you sure?')) {
+            let response = axios.delete(baseURL + '/deleteProfile/' + this.state.user_id)
+            console.log(response)
+            axios.delete(baseURL + '/deleteUsername/' + this.state.user_id)
+            window.location = '/'
+
+        }
+
+
 
     }
 }
